defconfig: 8226: Set ROW as the default block I/O scheduler
The ROW block I/O scheduler was developed with the mobile devices needs
in mind, where we favor user experience upon everything else, thus we
want to give READ I/O requests as much priority as possible.

The main idea of the ROW scheduling policy is to dispatch READ requests
with higher priority, while write starvation is considered.

The benefit of ROW block I/O scheduler is to improve the READ throughput
and latency in READ/WRITE collisions.
The READ latency worst case is improved both in sequential and random read
scenarios. For example, in random tiotest the worst READ latency
is decreased from 3940 msec with CFQ to 1180 msec with ROW.

In READ/WRITE collisions, the lmdd tests showed READ throughput
improvement from 32.62 MB/s with CFQ to 106.86 MB/s with ROW, with no
degradation to the WRITE throughput. The READ latency histogram shows
that when ROW is the current scheduler, the READ latency of 95% of the
requests fall under latency of less than 10 msec. The histogram for the
same scenario with CFQ as the current scheduler shows that 95% of the
READ requests are spread around latency of 10-40 msec.

Change-Id: I55415c1cd4bbb6f9c66d5d3903f55145b82ed53b
Signed-off-by: Tatyana Brokhman <tlinder@codeaurora.org>